Move Apache classes in a more standard package
